refactor: Move printing util to print file
